Simplified library structure and add documentation
