+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "sbf"
-version = "0.1.0"
+version = "0.1.1"
 authors = ["Filippo Berto <berto.f@protonmail.com>"]
 edition = "2018"
 homepage = "https://github.com/bertof/sbf-rs"
